Tidy up project structure
