+++ generate/full.py
