+++ pretrain/shakespeare.py
@@ -2,7 +2,8 @@
 This script is a placeholder for training LLaMA from scratch.
 Currently, it just trains on the Shakespeare dataset.
 """
-
+from pathlib import Path
+import sys
 import os
 import time
 from functools import partial
@@ -15,6 +16,10 @@ import torch
 from torch.distributed.fsdp.wrap import transformer_auto_wrap_policy
 
 import numpy as np
+
+# support running without installing as a package
+wd = Path(__file__).parent.parent.resolve()
+sys.path.append(str(wd))
 
 from lit_llama.model import Block, LLaMA, LLaMAConfig
 from lit_llama.utils import save_model_checkpoint
